Restructure services
