+++ src/polyswarmd/services/artifact/__init__.py
@@ -1,0 +1,2 @@
+from .client import AbstractArtifactServiceClient
+from .service import ArtifactServices
