+++ src/polyswarmd/services/consul.py
