+++ src/polyswarmd/services/ethereum/__init__.py
@@ -1,0 +1,1 @@
+from .service import EthereumService
